% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decluster.R
\name{decluster}
\alias{decluster}
\title{Decluster a Time Series}
\usage{
decluster(complete_series, obs_times = NULL)
}
\arguments{
\item{complete_series}{(numeric vector) The time series.}

\item{obs_times}{(numeric vector or NULL) If NULL, ignored; otherwise, the
observed times of the cluster maximums are returned too.}
}
\value{
An S3 object of class \code{declustered_series} with components

  \describe{

  \item{\code{declustered_series}}{The cluster maximums.}

  \item{\code{declustered_times}}{If \code{obs_times} is non NULL, the
  observed times of the cluster maximums; otherwise NULL.}

  }
}
\description{
Decluster a time series by forming clusters and returning only
  cluster maximums
}
\details{
Clusters are formed by sequential observations above the series mean
  value.  All observations below the mean value are discarded.  Cluster
  maximums are returned.
}
\examples{

complete_series <- -jp1tap1715wind270$value

declustered_obs <- decluster(complete_series)

}

