% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{fullAnalysis}
\alias{fullAnalysis}
\title{Full Analysis}
\usage{
fullAnalysis(complete_series, length_series, n_min, n_max, length_target_series,
  n_starts = 20, wplot_filename = NULL, BW = NULL, n_mc = 1000,
  max_dist_hist_filename = NULL, n_boot = 1000, ci_level = 0.8)
}
\arguments{
\item{complete_series}{(numeric vector) The observed time series that is a
realization of the random process.}

\item{length_series}{(numeric scalar) The length of the time series in units
of time (seconds, minutes, hours, etc.).}

\item{n_min}{(numeric scalar) The minimum number of observations to include
in the search for the optimal threshold.}

\item{length_target_series}{(numeric vector) The distribution of the peak
depends on the length of observation, which could be different from the
length of the original series.  Specify one or more lengths in units of
time (seconds, minutes, hours, etc.) for the length of series for which the
distribution of the peak is sought.  The units of time should be the same
as for the argument \code{length_series}.}

\item{n_starts}{(numeric scalar) The number of times to restart the
optimization algorithm from a random starting location; the default is 20.}

\item{wplot_filename}{(NULL or string) If NULL the plot is not created; else,
it is saved to "wplot_filename.pdf"; the default is NULL.}

\item{BW}{(logical) Create the Wplot in black and white (TRUE) or color
(FALSE); the default is TRUE.}

\item{n_mc}{(numeric scalar) The number of Monte Carlo samples from which to
construct the distribution of the peak; the defaults is 1000.}

\item{max_dist_hist_filename}{(NULL or string) If NULL the plot is not
created; else, it is saved to
"max_dist_hist_filename_length_target_series.pdf"; the default is NULL.}

\item{n_boot}{(numeric scalar) The number of bootstrap samples for the
uncertainty evaluation; the default is 1000.}

\item{ci_level}{(numeric scalar) The level of the confidence interval for the
mean of the peak distribution; the default is 0.8.}

\item{m_max}{(numeric scalar) The maximum number of observations to include
in the search for the optimal threshold.}
}
\value{
Returned invisibly, a data frame with four columns.  The first column
  is the estimated mean of the distribution of the peak.  The second column
  is a bootstrap standard error for the mean, and the third and fourth
  columns provide a percentile bootstrap confidence interval.
}
\description{
Estimates and characterizes uncertainty for the distribution of
  the peak value of a random process using the full peaks over threshold
  model.
}
\details{
See the vignette by running \code{vignette(topic = 'potMax_details',
  package = 'potMax')}.
}
\examples{

\dontrun{
  complete_series <- -jp1tap1715wind270$value
  fullAnalysis(complete_series = complete_series,
               length_series = 100,
               n_min = 10,
               n_max = 100,
               length_target_series = c(100, 167),
               wplot_filename = './tap1715_dir270_full_wplot',
               max_dist_hist_filename = './tap1716_dir270_full_max_dist_hist')

}

}

