% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_search.R
\name{fullEstThreshold}
\alias{fullEstThreshold}
\title{fullEstThreshold}
\usage{
fullEstThreshold(x, lt, n_min, n_max, n_starts, progress_tf = TRUE)
}
\arguments{
\item{lt}{The lenght of time in seconds over which the observations are
recorded}

\item{n_min}{The minimum number of thresholded observations to include}

\item{n_max}{The maximum number of thresholded observations to include}

\item{y_all}{The unthresholded series of observations}
}
\description{
fullEstThreshold
}
\details{
Select a threshold when using the full POT model
}
