% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMLE.R
\name{fullMLE}
\alias{fullMLE}
\alias{fullMLE.thresholded_series}
\alias{fullMLE.default}
\title{Maximum Likelihood Estimation for the Full Model}
\usage{
fullMLE(x, n_starts, hessian_tf, ...)

\method{fullMLE}{thresholded_series}(x, n_starts, hessian_tf)

\method{fullMLE}{default}(x, lt, thresh, n_starts, hessian_tf)
}
\arguments{
\item{x}{An S3 object of class \code{thresholded_series} or a numeric vector.
If the latter, the values used in fitting.}

\item{n_starts}{(numeric scalar) The number of random starts to use in the
search for the maximum}

\item{hessian_tf}{(logical scalar) Compute the Hessian matrix (TRUE) or not}

\item{lt}{(numeric scalar) The length of time over which data were observed
in units of time (seconds, minutes, hours, etc.)}

\item{thresh}{(numeric scalar) The threshold}
}
\value{
An S3 object of class \code{full_pot_fit}, which contains the
  estimated parameters \code{$par}, the threshold \code{$thresh}, the Hessian
  matrix \code{$lhessian} if requested, and the data used for the fit
  \code{$x}.
}
\description{
Maximizes the 2D extremal Poisson process likelihood that uses
  the full intensity function
}
\details{
The likelihood is

\deqn{\Big(\prod_{i = 1}^I \lambda(t_i,
y_i)\Big)\exp\Big[-\int_\mathcal{D} \lambda(t, y)dtdy\Big]}

where

\deqn{\lambda(t, y) = \frac{1}{\sigma}\Big[1 + \frac{k(y -
\mu)}{\sigma}\Big]^{-1/k - 1}_+}
}
\section{Methods (by class)}{
\itemize{
\item \code{thresholded_series}: 

\item \code{default}: 
}}

\examples{

\dontrun{
complete_series <- -jp1tap1715wind270$value

declustered_obs <- decluster(complete_series)

thresholded_obs <- fullEstThreshold(x = declustered_obs,
                                    lt = 100,
                                    n_min = 10,
                                    n_max = 100)

full_pot_fit <- fullMLE(x = thresholded_obs,
                        hessian_tf = TRUE)
}

}
