% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_year_vals.R
\name{fullNYear}
\alias{fullNYear}
\title{fullNYear}
\usage{
fullNYear(x, N, ...)
}
\arguments{
\item{x}{}

\item{N}{The N in N-year return value}
}
\description{
fullNYear
}
\details{
Calculates the N-year return value from a fitted POT model with
  a non-zero tail length parameter
}
