% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_search.R
\name{gumbelEstThreshold}
\alias{gumbelEstThreshold}
\title{gumbelEstThreshold}
\usage{
gumbelEstThreshold(x, lt, n_min, n_max)
}
\arguments{
\item{lt}{The lenght of time in seconds over which the observations are
recorded}

\item{n_min}{The minimum number of thresholded observations to include}

\item{n_max}{The maximum number of thresholded observations to include}

\item{y_all}{The unthresholded series of observations}
}
\description{
gumbelEstThreshold
}
\details{
Select a threshold when using the Gumble POT modle
}
