% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_dist_multi_thresh_fits.R
\name{gumbelMaxDist.gumbel_multi_fit}
\alias{gumbelMaxDist.gumbel_multi_fit}
\title{Distribution of the Maximum Using the Gumbel Model with Many Thresholds}
\usage{
\method{gumbelMaxDist}{gumbel_multi_fit}(x, lt_gen, n_mc,
  progress_tf = TRUE)
}
\arguments{
\item{x}{An S3 object of class \code{gumbel_multi_fit}.}

\item{lt_gen}{(numeric scalar) Length of each generated series.  The units
(seconds, minutes, hours, etc.) should be consistent with the value of
\code{lt} provided to \code{gumbelMultiFit}.}

\item{n_mc}{(numeric scalar) The number of samples to draw from the mixture
distribution of the maximums}

\item{progress_tf}{(logical scalar) Display a progress bar if TRUE, else not.}
}
\value{
An S3 object of class \code{gumbel_max_dist_multi_thresh} with elements

\describe{
  \item{\code{$mu}}{The estimated location parameter for each threshold}

  \item{\code{$sigma}}{The estimated scale parameter for each threshold}

  \item{\code{$thres}}{The thresholds}

  \item{\code{$lt_gen}}{The value of the \code{lt_gen} argument}

  \item{\code{$n_each}}{The number of samples drawn from each mixture
  component.  The sum of \code{$n_each} should be \code{n_mc}}

  \item{\code{$max_dist}}{A numeric vector of length \code{n_mc} containing
  the samples from the mixture distribution of the maximums}

}
}
\description{
Empirically build the distribution of the maximum value over
  some user defined length of time assuming the underlying data generating
  mechanism is a 2D extremal Poisson process with the Gumbel like intensity
  function (i.e. the tail length is zero), but without specifying a single
  threshold.
}
\details{
The results of fitting a Gumbel like 2D extremal Poisson process for
  many thresholds are fed into this function.  Random processes are
  repeatedly generated from each fitted model, according to the weights
  described in \code{gumbelMultiFit}, and the maximum of each random process
  is recorded.  The recoreded maximums represent an iid sample from a mixture
  of the distributions of maximum values.  Each potential threshold gives
  rise to a different distribution of the maximum value. Note that the
  desired length of the processes generated can be different from the length
  of time over which the data used to fit the models were observed.
}
\examples{

\dontrun{

complete_series <- -jp1tap1715wind270$value

declustered_obs <- decluster(complete_series)

gumbel_multi_fit <- gumbelMultiFit(x = declustered_obs, lt = 100,
                                   n_min = 10, n_max = 50,
                                   weight_scale = 5)

gumbel_max_dist <- gumbelMaxDist(x = gumbel_multi_fit, lt_gen = 200, n_mc = 1000)
}

}
