% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_max_dist.R
\name{gumbelMaxDist}
\alias{gumbelMaxDist}
\title{gumbelMaxDist}
\usage{
gumbelMaxDist(x, lt_gen, n_mc, ...)
}
\arguments{
\item{lt_gen}{Lengt of each generated series in seconds}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter}

\item{thresh}{The threshold}
}
\description{
gumbelMaxDist
}
\details{
Generates the empirical distribution of the maximum value from a
  Gumbel (zero tail length) POT model over a specified length of time
}

