% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_dist_multi_thresh_fits.R
\name{gumbelMaxDistUncert.gumbel_multi_fit}
\alias{gumbelMaxDistUncert.gumbel_multi_fit}
\title{Uncertainty in the Distribution of the Maximum Using the Gumbel Model
  With Many Thresholds}
\usage{
\method{gumbelMaxDistUncert}{gumbel_multi_fit}(x, declust_obs, lt_gen,
  n_mc, n_boot, progress_tf = TRUE)
}
\arguments{
\item{x}{An S3 object of class \code{gumbel_multi_fit}.}

\item{declust_obs}{(numeric vector) The observed data used by
\code{gumbelMultiFit}.  This will very often be the
\code{$declustered_series} element of an S3 object of class
\code{declustered_series}}

\item{lt_gen}{(numeric scalar) Length of each generated series.  The units
(seconds, minutes, hours, etc.) should be consistent with the value of
\code{lt} provided to \code{gumbelMLE}.}

\item{n_mc}{(numeric scalar) The number of samples to draw from the
distribution of the maximum}

\item{n_boot}{(numeric scalar) The number of bootstrap replicates of the
distribution of the maximum to create.}

\item{progress_tf}{Display a progress bar if TRUE, else not.}
}
\value{
An S3 object of class \code{gumbel_max_dist_uncert_multi_thresh},
  which is a list of length \code{n_boot} of S3 objects of class
  \code{gumbel_max_dist_multi_thresh}.
}
\description{
Evaluate uncertainty in the mixture of distributions of maximums
  assuming the underlying data generating mechanism is the 2D extremal
  Poisson process with the Gumbel like intensity function (i.e. the tail
  length is zero), but without specifying a single threshold.  The mixture is
  across multiple thresholds.
}
\details{
The results of fitting a many Gumbel like 2D extremal Poisson
  process are fed into this function.  The declustered data are repeatedly
  sampled with replacement, and for each resampled data set the distribution
  of the maximum is empirically constructed as described in
  \code{gumbelMaxDist.gumbel_multi_fit}. The bootstrap replicates of the
  mixture of distributions of the maximum may be used to quantify uncertainty
  and construct intervals.
}
\examples{

\dontrun{

complete_series <- -jp1tap1715wind270$value

declustered_obs <- decluster(complete_series)

gumbel_multi_fit <- gumbelMultiFit(x = declustered_obs, lt = 100,
                                   n_min = 10, n_max = 50,
                                   weight_scale = 5)

gumbel_multi_fit_uncert <- gumbelMaxDistUncert(x = gumbel_multi_fit,
                                               declust_obs = declustered_obs$declustered_series,
                                               lt_gen = 200,
                                               n_mc = 1000,
                                               n_boot = 200)
}

}
