% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbelMLE.R
\name{gumbelMLE}
\alias{gumbelMLE}
\alias{gumbelMLE.thresholded_series}
\alias{gumbelMLE.default}
\title{Maximum Likelihood Estimation for the Gumble Model}
\usage{
gumbelMLE(x, hessian_tf, ...)

\method{gumbelMLE}{thresholded_series}(x, hessian_tf)

\method{gumbelMLE}{default}(x, lt, thresh, hessian_tf)
}
\arguments{
\item{x}{An S3 object of class \code{thresholded_series} or a numeric vector.
If the latter, the values used in fitting.}

\item{hessian_tf}{(logical scalar) Compute the Hessian matrix (TRUE) or not.}

\item{lt}{(numeric scalar) The length of the time series in units
of time (seconds, minutes, hours, etc.).}

\item{thresh}{(numeric scalar) The threshold for the values}
}
\value{
An S3 object of class \code{gumbel_pot_fit} with elements

\describe{

  \item{\code{$par}}{(numeric vector of length 2) The estimated location and
  scale parameter, respectively}

  \item{\code{$lhessian}}{The Hessian matrix at the MLE if requested, else NULL}

  \item{\code{$y}}{The observed values used to fit the model}

  \item{\code{$thresh}}{The threshold}
}
}
\description{
Solves the score equations for the 2D extremal Poisson process
  likelihood using the Gumbel like intensity function
}
\details{
The likelihood is

  \deqn{\Big(\prod_{i = 1}^I \lambda(t_i,
  y_i)\Big)\exp\Big[-\int_\mathcal{D} \lambda(t, y)dtdy\Big]}

  where

  \deqn{\lambda(t, y) = \frac{1}{\sigma}\exp\Big[\frac{-(y -
  \mu)}{\sigma}\Big]}
}
\section{Methods (by class)}{
\itemize{
\item \code{thresholded_series}: 

\item \code{default}: 
}}

\examples{

\dontrun{

complete_series <- -jp1tap1715wind270$value

declustered_obs <- decluster(complete_series)

thresholded_obs <- gumbelEstThreshold(x = declustered_obs,
                                      lt = 100,
                                      n_min = 10,
                                      n_max = 100)

gumbel_pot_fit <- gumbelMLE(x = thresholded_obs,
                            hessian_tf = TRUE)
}

}
