% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_threshold_fits.R
\name{gumbelMultiFit}
\alias{gumbelMultiFit}
\title{gumbelMultiFit}
\usage{
gumbelMultiFit(x, lt, n_min, n_max, weight_scale, progress_tf = TRUE)
}
\arguments{
\item{lt}{The lenght of time in seconds over which the observations are
recorded}

\item{n_min}{The minimum number of thresholded observations to include}

\item{n_max}{The maximum number of thresholded observations to include}

\item{weight_scale}{The weights are
exp[-weight_scale/(max_w - min_w)*(w - min_w)]
normalized to sum to unity}

\item{y_all}{The unthresholded series of observations}
}
\description{
gumbelMultiFit
}
\details{
Fit the Gumble POT model with multiple thresholds
}
