% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_year_vals.R
\name{gumbelNYear}
\alias{gumbelNYear}
\title{gumbelNYear}
\usage{
gumbelNYear(x, N, ...)
}
\arguments{
\item{x}{}

\item{N}{The N in N-year return value}
}
\description{
gumbelNYear
}
\details{
Calculates the N-year return value from a fitted POT model with the
  Gumbel tail length parameter, i.e., 0
}
