% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_year_vals.R
\name{gumbelNYearUncert}
\alias{gumbelNYearUncert}
\alias{gumbelNYearUncert.gumbel_pot_fit}
\alias{gumbelNYearUncert.default}
\title{Uncertainty in Return Values for the Gumbel Model}
\usage{
gumbelNYearUncert(x, N, n_boot, ...)

\method{gumbelNYearUncert}{gumbel_pot_fit}(x, N, n_boot)

\method{gumbelNYearUncert}{default}(x, cov_mat, thresh, N, n_boot)
}
\arguments{
\item{x}{An S3 object of type \code{gumbel_pot_fit} or a numeric vector of
length 2.  If the latter, the first element is interpreted as the location
parameter \eqn{\mu}, and the second element is interpreted as the scale
parameter \eqn{\sigma}.}

\item{N}{(numeric scalar) The N in N-year return value.  This is a bit of a
misnomer since the unit of time does not have to be years.  The function
can calculate N-second, N-minute, N-hour, etc. return values as well.  In
fact, the unit of time is the same unit of time passed in for the
\code{lt} argument of \code{gumbelMLE}.  Naming the function for the unit
of time year is simply due to my past experince with calculating return
values on the time scale of years.}

\item{n_boot}{(numeric scalar) The number of bootstrap samples}

\item{cov_mat}{The covariance matrix used to perturb the estimated
parameters.  This will most usually be the negative inverse of the Hessian
matrix at the MLE}
}
\value{
An S3 object of class \code{gumbel_N_year_val_uncert} with elements

\describe{

  \item{\code{$par}}{numeric vector of length 2 containing the location and
  scale parameters, respectively of the 2D extremal Poisson process used to
  calculat the return value}

  \item{\code{$thresh}}{The threshold}

  \item{\code{$N}}{The value inputted for \code{N}}

  \item{\code{$boot_samps}}{(numeric vector of length \code{n_boot}) The
  bootstrap sample of return values}
}
}
\description{
The function produces a bootstrap sample of return values using
  the Gumbel like 2D extremal Poisson process
}
\details{
Repeatedly solves the equation

\deqn{\int_{y_N}^\infty\int_0^1\lambda(t, y)dtdy = \frac{1}{N}}

where \eqn{\lambda(t, y)} is given in the documentation for \code{gumbelMLE}
for perturbed values of \eqn{\mu} and \eqn{\sigma}.  The perturbed values are
obtained by using the Hessian matrix and multivariate Guassian distribution
to perturb the MLE.
}
\section{Methods (by class)}{
\itemize{
\item \code{gumbel_pot_fit}: 

\item \code{default}: 
}}

\examples{

\dontrun{

complete_series <- -jp1tap1715wind270$value

declustered_obs <- decluster(complete_series)

thresholded_obs <- gumbelEstThreshold(x = declustered_obs,
                                      lt = 100,
                                      n_min = 10,
                                      n_max = 100)

gumbel_pot_fit <- gumbelMLE(x = thresholded_obs,
                            hessian_tf = TRUE)

500_second_val_uncert <- gumbelNYearUncert(x = gumbel_pot_fit, N = 500,
                                           n_boot = 1000)
}

}
