% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_search.R
\name{gumbelWPlot}
\alias{gumbelWPlot}
\title{gumbelWPlot}
\usage{
gumbelWPlot(x, tf_plot, BW, details, ...)
}
\arguments{
\item{tf_plot}{}

\item{BW}{(logical) only use black and white for the plot.  Ignored if
tf_plot == FALSE}

\item{details}{}

\item{...}{}

\item{mu}{}

\item{sigma}{}

\item{y}{}

\item{thresh}{}
}
\description{
gumbelWPlot
}
\details{
this is the W plot from the 2004 Richard Smith book chapter
}
