% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pot_poisson_process.R
\name{rFullPOT}
\alias{rFullPOT}
\title{Random Sample from the Full Poisson Process}
\usage{
rFullPOT(mu, sigma, k, thresh, lt)
}
\arguments{
\item{mu}{(numeric scalar) Location parameter}

\item{sigma}{(numeric scalar) Scale parameter}

\item{k}{(numeric scalar) Tail length parameter}

\item{thresh}{(numeric scalar) The threshold}

\item{lt}{(numeric scalar) Lengt of the series in units of time (seconds,
minutes, hours, etc.)}
}
\value{
A list with components \code{$t} the times of the threshold crossings
  in the same units of \code{lt} and \code{$y} the value of the threshold
  crossings (NOT exceedances of/differences from the threshold)
}
\description{
Generate a random sample from the 2D Poisson process with the
  full intensity function
}
\details{
The intensity function is described in \code{fullMLE}
}
\examples{

\dontrun{

f_pot_samp <- rFullPOT(mu = 50, sigma = 5, k = 0.2, thresh = 55, lt = 100)

}

}
