% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pot_poisson_process.R
\name{rGumbelPOT}
\alias{rGumbelPOT}
\title{Random Sample from the Gumbel like Poisson Process}
\usage{
rGumbelPOT(mu, sigma, thresh, lt)
}
\arguments{
\item{mu}{(numeric scalar) Location parameter}

\item{sigma}{(numeric scalar) Scale parameter}

\item{thresh}{(numeric scalar) The threshold}

\item{lt}{(numeric scalar) Lengt of the series in units of time (seconds,
minutes, hours, etc.)}
}
\value{
A list with components \code{$t} the times of the threshold crossings
  in the same units of \code{lt} and \code{$y} the value of the threshold
  crossings (NOT exceedances of/differences from the threshold)
}
\description{
Generate a random sample from the 2D Poisson process with a
  Gumbel like intensity function
}
\details{
The intensity function is described in \code{gumbelMLE}
}
\examples{

\dontrun{

g_pot_samp <- rGumbelPOT(mu = 50, sigma = 5, thresh = 55, lt = 100)

}

}
