% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{threshPlot}
\alias{threshPlot}
\alias{threshPlot.thresholded_series}
\alias{threshPlot.gumbel_multi_fit}
\alias{threshPlot.full_multi_fit}
\title{Plot Fit Statistics versus Thresholds}
\usage{
threshPlot(x, ...)

\method{threshPlot}{thresholded_series}(x, ...)

\method{threshPlot}{gumbel_multi_fit}(x, ...)

\method{threshPlot}{full_multi_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{thresholded_series},
\code{gumbel_multi_fit}, or \code{full_multi_fit}}
}
\value{
A \code{ggplot} object
}
\description{
This function creates a plot of of the fit statistics versus the
  associated thresholds
}
\details{
The fit statistics are the maximum vertical distances of the points
  to the \eqn{45^\circ} line in the W-plots described in \code{gumbelWPlot}
  and \code{fullWPlot}
}
\section{Methods (by class)}{
\itemize{
\item \code{thresholded_series}: 

\item \code{gumbel_multi_fit}: 

\item \code{full_multi_fit}: 
}}

\examples{

\dontrun{
  complete_series <- -jp1tap1715wind270$value

  declustered_obs <- decluster(complete_series)

  thresholded_series <- gumbelEstThreshold(x = declustered_obs,
                                           lt = 100,
                                           n_min = 10,
                                           n_max = 100)

  threshPlot(thresholded_series)
}

}
