% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_max_dist.R
\name{fullMaxDist}
\alias{fullMaxDist}
\title{fullMaxDist}
\usage{
fullMaxDist(x, lt_gen, n_mc, progress_tf = TRUE, ...)
}
\arguments{
\item{lt_gen}{Lengt of each generated series in seconds}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter}

\item{k}{Tail length parameter}

\item{thresh}{The threshold}
}
\description{
fullMaxDist
}
\details{
Generates the empirical distribution of the maximum value from a
  full (non-zero tail length) POT model over a specified length of time
}
