% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMLE.R
\name{fullMLE}
\alias{fullMLE}
\alias{fullMLE.default}
\alias{fullMLE.thresholded_series}
\title{Maximum Likelihood Estimation for the Full Model}
\usage{
fullMLE(x, n_starts, hessian_tf, ...)

\method{fullMLE}{thresholded_series}(x, n_starts, hessian_tf)

\method{fullMLE}{default}(x, lt, thresh, n_starts, hessian_tf)
}
\arguments{
\item{x}{An S3 object of class \code{thresholded_series} or a numeric vector.
If the latter, the values used in fitting.}

\item{n_starts}{(numeric scalar) The number of random starts to use in the
search for the maximum}

\item{lt}{(numeric scalar) The length of time over which data were observed
in seconds}

\item{thresh}{(numeric scalar) The threshold}

\item{N}{- (numeric scalar) The number of observations that exceed the
threshold (also the length of y)}

\item{hessian}{(logical scalar) Compute the Hessian matrix (TRUE) or not}
}
\description{
Maximizes the Poisson process likelihood that uses the full
  intensity function
}
\details{
The likelihood is

\deqn{\left(\prod_{i = 1}^I \lambda(t_i,
y_i)\right)\exp\left[-\int_\mathcal{D} \lambda(t, y)dtdy\right]}

where

\deqn{\lambda(t, y) = \frac{1}{\sigma}\left[1 + \frac{k(y -
\mu)}{\sigma}\right]^{-1/k - 1}_+}
}
\section{Methods (by class)}{
\itemize{
\item \code{thresholded_series}: 

\item \code{default}: 
}}
\examples{

\dontrun{
complete_series <- -jp1tap1715wind270$value

declustered_obs <- decluster(complete_series)

thresholded_obs <- fullEstThreshold(x = declustered_obs,
                                    lt = 100,
                                    n_min = 10,
                                    n_max = 100)

full_pot_fit <- fullMLE(x = thresholded_obs,
                        hessian_tf = TRUE)
}

}

