% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_threshold_fits.R
\name{fullMultiFit}
\alias{fullMultiFit}
\title{fullMultiFit}
\usage{
fullMultiFit(x, lt, n_min, n_max, weight_scale, n_starts)
}
\arguments{
\item{lt}{The lenght of time in seconds over which the observations
are recorded}

\item{n_min}{The minimum number of thresholded observations to
include}

\item{n_max}{The maximum number of thresholded observations to
include}

\item{weight_scale}{The weights are
exp[-weight_scale/(max_w - min_w)*(w - min_w)]
normalized to sum to unity}

\item{n_starts}{(numeric scalar) The number of random starts to use
in the search for the maximum}

\item{y_all}{The unthresholded series of observations}
}
\description{
fullMultiFit
}
\details{
Fit the full POT model with multiple thresholds
}
