% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_max_dist.R
\name{gumbelMaxDistUncert}
\alias{gumbelMaxDistUncert}
\title{gumbelMaxDistUncert}
\usage{
gumbelMaxDistUncert(x, lt_gen, n_mc, n_boot, ...)
}
\arguments{
\item{lt_gen}{Lengt of each generated series in seconds that the maximum is
take over}

\item{n_mc}{The number of Monte Carlo samples from which the distribution of
the maximum is built}

\item{n_boot}{The number of bootstrap samples to base uncertainties on}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter}

\item{cov_mat}{The estimated covariance matrix for the estimation of mu and
sigma}

\item{thresh}{The threshold}
}
\description{
gumbelMaxDistUncert
}
\details{
Generates bootstrap samples of the empirical distribution of the
  maximum value from a Gumbel (zero tail length) POT model over a specified
  #'   length of time
}

